Add matomo metrics integration
This patch adds a configuration to use a matomo server for metrics
instead of eosmetrics. This configuration can be used for flathub
version just adding it to the manifest.json in the clubhouse
config-opts:

     "-Deosmetrics=False",

This patch also includes a queue for events, so we can store events if
there's no connection or if the matomo server is down or something like
that. This queue will be dequeued once an event is sent correctly.

To anonimize the data, a random unique user id is generated. This id is
stored on filesystem so we can use the same one everytime and we can
identify the same user without getting any real information about the
user.

https://phabricator.endlessm.com/T30942