metrics: Add url to matomo metrics
This patch adds the url param to the matomo track information. The url
is a base attribute in matomo because it's mostly used to track
webpages. To adapt to our generic data, this patch adds a new function
that converts the event information in a real url so we can see this
information in a clear way in the matomo web interface.

This url will be useful to track easily what users are doing with the
app and even to do create custom queries and aggregate this information.

https://phabricator.endlessm.com/T30942